+++ packages/coin98-wallet/src/lib/coin98-wallet.ts
@@ -12,7 +12,7 @@ import type {
 import { getActiveAccount } from "@near-wallet-selector/core";
 import type { InjectedCoin98 } from "./injected-coin98-wallet";
 import { signTransactions } from "@near-wallet-selector/wallet-utils";
-import type { FinalExecutionOutcome } from "near-api-js/lib/providers";
+import type { FinalExecutionOutcome } from "near-api-js/lib/providers/index.js";
 import * as nearAPI from "near-api-js";
 import icon from "./icon";