+++ packages/coin98-wallet/src/lib/injected-coin98-wallet.ts
@@ -2,7 +2,7 @@ import type {
   SignedMessage,
   SignMessageParams,
 } from "@near-wallet-selector/core";
-import type { Signer } from "near-api-js/lib/signer";
+import type { Signer } from "near-api-js/lib/signer.js";
 
 interface IConnectParams {
   prefix: string;