+++ packages/core/src/lib/wallet-selector.ts
@@ -92,7 +92,7 @@ export const setupWalletSelector = async (
 
   const emitter = new EventEmitter<WalletSelectorEvents>();
   const store = await createStore(storage);
-  const network = await getNetworkPreset(
+  const network = getNetworkPreset(
     options.network.networkId as NetworkId,
     params.fallbackRpcUrls
   );
@@ -111,7 +111,8 @@ export const setupWalletSelector = async (
     provider: new Provider(rpcProviderUrls),
   });
 
-  await walletModules.setup();
+  await walletModules.setupWalletModules();
+  await walletModules.setupStorage();
 
   if (params.allowMultipleSelectors) {
     return createSelector(options, store, walletModules, emitter);