+++ packages/ethereum-wallets/src/lib/index.ts
@@ -4,10 +4,10 @@ import type {
   ExecutionStatus,
   FinalExecutionOutcome,
   FunctionCallPermissionView,
-} from "near-api-js/lib/providers/provider";
-import { JsonRpcProvider } from "near-api-js/lib/providers";
-import { stringifyJsonOrBytes } from "near-api-js/lib/transaction";
-import { parseRpcError } from "near-api-js/lib/utils/rpc_errors";
+} from "near-api-js/lib/providers/provider.js";
+import { JsonRpcProvider } from "near-api-js/lib/providers/index.js";
+import { stringifyJsonOrBytes } from "near-api-js/lib/transaction.js";
+import { parseRpcError } from "near-api-js/lib/utils/rpc_errors.js";
 import {
   type WalletModuleFactory,
   type WalletBehaviourFactory,
@@ -56,7 +56,11 @@ const importBannedNearAddressesPackage = async () => {
 };
 
 import icon from "./icon";
-import { createTxModal, createChainSwitchModal } from "./modal";
+import {
+  createTxModal,
+  createChainSwitchModal,
+  createMessageModal,
+} from "./modal";
 import {
   ETHEREUM_ACCOUNT_ABI,
   DEFAULT_ACCESS_KEY_ALLOWANCE,
@@ -545,6 +549,63 @@ const EthereumWallets: WalletBehaviourFactory<
     transactions: Array<Optional<Transaction, "signerId" | "receiverId">>
   ) => {
     const nearTxs = await transformTransactions(transactions);
+    if (bannedNearAddressesPackage === null) {
+      await importBannedNearAddressesPackage();
+    }
+
+    let restrictedActionError = "";
+    for (let i = 0; i < nearTxs.length; i++) {
+      const tx = nearTxs[i];
+      for (let y = 0; y < tx.actions.length; y++) {
+        try {
+          const action = tx.actions[y];
+          let accountId = null;
+          if (action.type === "Transfer") {
+            accountId = tx.receiverId;
+          } else if (action.type === "FunctionCall") {
+            if (action.params.methodName.includes("transfer")) {
+              //@ts-ignore
+              accountId = action.params?.args?.receiver_id;
+            } else if (action.params.methodName === "storage_deposit") {
+              //@ts-ignore
+              accountId = action.params?.args?.account_id;
+            }
+          }
+
+          if (
+            accountId &&
+            bannedNearAddressesPackage?.isBannedNearAddress(
+              accountId.toLowerCase()
+            )
+          ) {
+            restrictedActionError = `Transferring funds to ${accountId} has been restricted due to security reasons in order to prevent users from losing funds. If you have any questions, feel free to contact NEAR Support through any official channel.`;
+            break;
+          }
+        } catch {
+          // eslint-disable-next-line no-console
+          console.error(
+            "Failed to validate NEAR transactions for transfers to restricted addresses."
+          );
+        }
+      }
+    }
+
+    if (restrictedActionError) {
+      await (() => {
+        return new Promise<void>((_resolve, reject) => {
+          const onCancel = () => {
+            reject(restrictedActionError);
+          };
+          const { showModal } = createMessageModal({
+            title: "Warning",
+            message: restrictedActionError,
+            onCancel,
+          });
+          showModal();
+        });
+      })();
+    }
+
     const [accountLogIn] = await getAccounts();
     // If transactions can be executed with FunctionCall access key do it, otherwise execute 1 by 1 with Ethereum wallet.
     if (accountLogIn.publicKey && nearTxs.length) {