+++ packages/nightly/src/lib/injected-nightly.ts
@@ -6,8 +6,8 @@ import type {
 import type {
   SignedTransaction as NearSignedTransaction,
   Transaction as NearTransaction,
-} from "near-api-js/lib/transaction";
-import type { PublicKey as NearPublicKey } from "near-api-js/lib/utils";
+} from "near-api-js/lib/transaction.js";
+import type { PublicKey as NearPublicKey } from "near-api-js/lib/utils/index.js";
 interface NightlyAccount {
   accountId: string;
   publicKey: NearPublicKey;