+++ packages/arepa-wallet/CHANGELOG.md
@@ -1,3 +1,11 @@
 # @near-wallet-selector/arepa-wallet
 
+## 9.3.0
+
+## 9.2.0
+
+### Patch Changes
+
+- [#1380](https://github.com/near/wallet-selector/pull/1380) [`b5bca7a`](https://github.com/near/wallet-selector/commit/b5bca7a66484686fad7c975b53b25fdd714421f5) Thanks [@gagdiez](https://github.com/gagdiez)! - fix: esm imports from near-api-js
+
 ## 9.1.0