+++ packages/ethereum-wallets/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @near-wallet-selector/ethereum-wallets
 
+## 9.3.0
+
+### Minor Changes
+
+- [#1382](https://github.com/near/wallet-selector/pull/1382) [`7c9bda8`](https://github.com/near/wallet-selector/commit/7c9bda8ee997b19d6d0f97b371e385c628019acb) Thanks [@yellowCrimsonGator](https://github.com/yellowCrimsonGator)! - Block Native Near transfers to restricted ETH addresses
+
+## 9.2.0
+
+### Minor Changes
+
+- [#1377](https://github.com/near/wallet-selector/pull/1377) [`9200528`](https://github.com/near/wallet-selector/commit/92005288d7ac8760cfca45365f8ae9e3bbcdfef1) Thanks [@yellowCrimsonGator](https://github.com/yellowCrimsonGator)! - Block tx to restricted adresses in eth wallet
+
+### Patch Changes
+
+- [#1380](https://github.com/near/wallet-selector/pull/1380) [`b5bca7a`](https://github.com/near/wallet-selector/commit/b5bca7a66484686fad7c975b53b25fdd714421f5) Thanks [@gagdiez](https://github.com/gagdiez)! - fix: esm imports from near-api-js
+
 ## 9.1.0
 
 ### Minor Changes