+++ packages/xdefi/CHANGELOG.md
@@ -1,3 +1,15 @@
 # @near-wallet-selector/xdefi
 
+## 9.3.0
+
+### Patch Changes
+
+- [#1383](https://github.com/near/wallet-selector/pull/1383) [`4123469`](https://github.com/near/wallet-selector/commit/4123469a1a23044896d1ee557e2a20c8ac3c04b8) Thanks [@ahnafalfariza](https://github.com/ahnafalfariza)! - Enhance load time by removing the waitFor function that checks if the wallet is installed
+
+## 9.2.0
+
+### Patch Changes
+
+- [#1380](https://github.com/near/wallet-selector/pull/1380) [`b5bca7a`](https://github.com/near/wallet-selector/commit/b5bca7a66484686fad7c975b53b25fdd714421f5) Thanks [@gagdiez](https://github.com/gagdiez)! - fix: esm imports from near-api-js
+
 ## 9.1.0